requirements.txt
qdrant-client
sentence-transformers
kubernetes
python-dotenv
groq
fastapi
uvicorn[standard]
websockets>=12.0
httptools
uvloop
